% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_frame.R
\name{as.track_frame}
\alias{as.track_frame}
\title{Convert an object to a Track Frame}
\usage{
as.track_frame(data, ...)
}
\arguments{
\item{data}{the object to be converted.}

\item{...}{additional arguments passed to the method.}
}
\value{
A track_frame object
}
\description{
Convert an object to a Track Frame
}
