% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_frame.R
\name{as.track_frame_from_cocomo}
\alias{as.track_frame_from_cocomo}
\title{Converts a cocomo object to a track_frame}
\usage{
as.track_frame_from_cocomo(xs, ys, timestamps, ids)
}
\arguments{
\item{xs}{matrix of x coordinates (UTM eastings) of all individuals in a group or population (rows) at every time point (columns) xs[i,t] gives the x / easting position of individual i at time point t}

\item{ys}{matrix of y coordinates (UTM northings) of all individuals in a group or population (rows) at every time point (columns) ys[i,t] gives the y / northing position of individual i at time point t}

\item{timestamps}{vector of timestamps in posixct corresponding to the columns of xs and ys matrices. Timestamps must be uniformly sampled, though it is possible to have gaps (e.g. between different days of recording)}

\item{ids}{data frame giving information about the tracked individuals, with rows correpsonding to the rows of the xs and ys matrices. There must be one column called id_code which contains a unique individual identifier for each animal (e.g. for meerkats: 'VCVM001', for hyenas: 'WRTH', for coatis: 'Luna') The other columns contained are flexible, and can include information on age, sex, dominance, etc}
}
\value{
an object of class track_frame
}
\description{
Converts a cocomo object to a track_frame
}
\examples{
#test data
padding <- 25
n_times <- 20 + 3 * padding
N <- 2

xs <- matrix(
  c(
    rep(-1, padding),
    c(-1, -0.9, -0.8, -0.7, -0.6, -0.5, -0.4, -0.3, -0.2, -0.1),
    rep(0, padding),
    c(0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1),
    rep(1, padding),
    rep(0, n_times)
  ),
  N,n_times,byrow=TRUE
)
ys <- matrix(
  rep(0, N*n_times),
  N,n_times,byrow=TRUE
)
timestamps <- as.POSIXct(1:n_times)

ids <- rbind.data.frame(list("id_code" = 'VCVM001', "age" = 10, "sex" = "m"),
                        list("id_code" = 'WRTH', "age" = 5, "sex" = "f"))
                        
tf <- as.track_frame_from_cocomo(xs, ys, timestamps, ids)
class(tf)
head(tf)
}
