% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackframe.R
\name{trackframe}
\alias{trackframe}
\alias{as.trackframe}
\alias{as.trackframe.data.frame}
\alias{as.trackframe.matrix}
\alias{as.trackframe.move2}
\alias{as.trackframe.sftrack}
\alias{as.trackframe.sf}
\alias{as.trackframe.trackframe}
\title{Convert an object to a \code{trackframe}}
\usage{
trackframe(
  data,
  time_col = tf_options("time_col"),
  easting_col = tf_options("easting_col"),
  northing_col = tf_options("northing_col"),
  id_col = tf_options("id_col"),
  sort = TRUE,
  coerce_to = "base",
  crs = NULL,
  ...
)

as.trackframe(
  data,
  time_col = tf_options("time_col"),
  easting_col = tf_options("easting_col"),
  northing_col = tf_options("northing_col"),
  id_col = tf_options("id_col"),
  sort = TRUE,
  coerce_to = "base",
  crs = NULL,
  ...
)

\method{as.trackframe}{data.frame}(
  data,
  time_col = tf_options("time_col"),
  easting_col = tf_options("easting_col"),
  northing_col = tf_options("northing_col"),
  id_col = tf_options("id_col"),
  sort = TRUE,
  coerce_to = "base",
  crs = NULL,
  ...
)

\method{as.trackframe}{matrix}(
  data,
  time_col = tf_options("time_col"),
  easting_col = tf_options("easting_col"),
  northing_col = tf_options("northing_col"),
  id_col = tf_options("id_col"),
  sort = TRUE,
  coerce_to = "base",
  crs = NULL,
  ...
)

\method{as.trackframe}{move2}(
  data,
  time_col = NULL,
  easting_col = NULL,
  northing_col = NULL,
  id_col = NULL,
  sort = TRUE,
  coerce_to = "base",
  ...
)

\method{as.trackframe}{sftrack}(
  data,
  time_col = NULL,
  easting_col = NULL,
  northing_col = NULL,
  id_col = NULL,
  sort = TRUE,
  coerce_to = "base",
  ...
)

\method{as.trackframe}{sf}(
  data,
  time_col = tf_options("time_col"),
  easting_col = NULL,
  northing_col = NULL,
  id_col = tf_options("id_col"),
  sort = TRUE,
  coerce_to = "base",
  ...
)

\method{as.trackframe}{trackframe}(
  data,
  time_col = NULL,
  easting_col = NULL,
  northing_col = NULL,
  id_col = NULL,
  sort = TRUE,
  coerce_to = NULL,
  crs = NULL,
  ...
)
}
\arguments{
\item{data}{a \code{data.frame}, \code{matrix}, \code{sftrack} object or \code{move2} object
containing timestamped coordinates of one or multiple animals' locations.
The object should have one row per location, and must contain at least a column for each of
time, easting, and northing.}

\item{time_col}{a character string or vector of strings specifying the name, or possible names, of the time
column in \code{data}.
If a vector is provided, the first element matching a column name of \code{data} is used.
The data in the time column must be of class \code{POSIXct} or coercible to \code{POSIXct}.}

\item{easting_col}{a character string or vector of strings specifying the name, or possible names, of the easting
column in \code{data}.
If a vector is provided, the first element matching a column name of \code{data} is used.
The data in the easting column must be numeric.}

\item{northing_col}{a character string or vector of strings specifying the name, or possible names, of the northing
column in \code{data}.
If a vector is provided, the first element matching a column name of \code{data} is used.
The data in the northing column must be numeric.}

\item{id_col}{an optional character string or vector of strings specifying the name, or possible names,
of the id column (the column containing the individual identifier(s) of each track) in \code{data}.
If a vector is provided, the first element matching a column name of \code{data} is used.
The data in the id column can be numeric, character, or factor.}

\item{sort}{logical, if true (default), data will be sorted by id_col and then by time_col.}

\item{coerce_to}{the type of dataframe that trackframe is coerced to.
\code{base}, \code{data.table} and \code{tibble} are supported.
The default is \code{base}, which coerces to a \code{data.frame} without \code{data.table} or \code{tbl} classes.
If NULL, the returned \code{trackframe} object takes the same dataframe type as the input \code{data}
object.}

\item{crs}{the numeric \href{https://epsg.io/}{EPSG coordinate reference system code} for the coordinates.
This argument is required for non-sf input.
Use \code{NA} to denote non-georeferenced coordinates.
The coordinate reference system must be Cartesian.}

\item{...}{Additional arguments (unused).}
}
\value{
A \code{trackframe} object with appropriate attributes set.
}
\description{
This function converts an object into a \code{trackframe} object, and
checks that the key columns exist and have valid data types.
The coordinate system must be Cartesian: either a projected coordinate system
for georeferenced data or another coordinate system in Euclidean space for non-georeferenced data
(e.g., for captive or simulated data).
}
\details{
When creating a \code{trackframe} object from another representation of animal track data,
\code{as.trackframe} identifies the columns representing
time (\code{time_col}), coordinates (\code{northing_col} & \code{easting_col}), and track id (\code{id_col}).
It obtains this information from 3 potential sources: 1) information inherited from the input
object (see below), 2) arguments supplied directly to \code{as.trackframe}, and/or 3) package-level
options(see \link{tf_options}).

If the input object is a \code{data.frame}, \code{data.table}, or \code{tibble},
no information about the (\code{time_col}), coordinates (\code{northing_col} & \code{easting_col}),
and track id (\code{id_col}) columns is inherited,
all are identified from other sources
(i.e., arguments supplied directly to the function, or package-level options).
If the input object is an \code{sf} object (but not \code{sftrack} or \code{move2}),
the \code{easting_col} and \code{northing_col} are identified from the input object's \code{geometry} column,
and the \code{time_col} and \code{id_col} are identified from other sources.
If the input object is an \code{sftrack} or \code{move2} object,
then all 4 of the key columns are identified from the input object's attributes.
See vignette \code{vignette("identifying-columns")} for details and examples.

Trackframe only supports Cartesian coordinates, i.e., coordinates that define locations on a
flat surface. Geographic coordinate systems, i.e., those that
define locations on a three-dimensional spherical model of the Earth using latitude and
longitude, are not supported by trackframe or by (most) functions using trackframe.
If you try to pass a non-Cartesian georeferenced coordinate reference system to
the \code{crs} argument, you will get an error. Data
containing latitude-longitude coordinates must first be projected before the data is
converted to a \code{trackframe} object.
If you set \code{crs = NA} for non-georeferenced coordinates, the function will assume that these are
Cartesian coordinates, though it has no way of checking or warning if not.
}
\examples{
library(trackframe)

# Create a dataframe
set.seed(2025)
df <- data.frame(
  x = rnorm(10),
  y = rnorm(10),
  t = 1:10,
  animal_id = c(rep('a', 5), rep('b', 5))
)

# Convert dataframe to trackframe, identify key columns by setting arguments
tf <- trackframe(df,
                 time_col = "t",
                 easting_col = "x",
                 northing_col = "y",
                 id_col = "animal_id",
                 crs = "EPSG:32632")

class(tf)
attributes(tf)

tf <- as.trackframe(df,
                    time_col = "t",
                    easting_col = "x",
                    northing_col = "y",
                    id_col = "animal_id",
                    crs = "EPSG:32632")

# Convert dataframe to trackframe, identify key columns by having as.trackframe match column
# names to tf_options
tf <- as.trackframe(df_mini, crs = NA)

# Convert dataframe to trackframe, identify key columns by having as.trackframe match column
# names to multiple options given in arguments
tf <- as.trackframe(df_mini,
                    time_col = c("t", "time", "timestamp"),
                    easting_col = c("x", "easting", "horiz"),
                    northing_col = c("y", "northing", "vert"),
                    id_col = c("animal", "id"),
                    crs = NA)

# Convert dataframe to trackframe, identify key columns by setting arguments, also set
# crs = NA to indicate non georeferenced coordinates
tf <-  as.trackframe(df,
                    time_col = "t",
                    easting_col = "x",
                    northing_col = "y",
                    id_col = "animal_id",
                    crs = NA)

# Convert dataframe to trackframe, identify key columns by having as.trackframe match column
# names to tf_options, set crs = NA to indicate non georeferenced coordinates, coerce output
# to be:
# ... a plain (base R) data.frame
tf_df <- as.trackframe(df, coerce_to = "base", crs = NA)
attributes(tf_df)$class
# ... a data.table
tf_dt <- as.trackframe(df, coerce_to = "data.table", crs = NA)
attributes(tf_dt)$class
# ... a tibble.
tf_tib <- as.trackframe(df, coerce_to = "tibble", crs = NA)
attributes(tf_tib)$class

# Projected CRS, acceptable
tf_df <- as.trackframe(df, coerce_to = "base", crs = "EPSG:32632")
# Non Cartesian crs, e.g. "EPSG:4326", is not acceptable

# Example for move2 objects

# Key columns and crs are identified from the attributes of the move2 object
tf <- as.trackframe(data = move2_mini)


# Example for sftrack objects

# Key columns and crs are identified from the attributes of the sftrack object
tf <- as.trackframe(data = sftrack_mini)


}
\seealso{
\link{tf_options}, \link{tf_backtransform}
}
