% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_suggestion.R
\name{calculate_utm_zone_crs}
\alias{calculate_utm_zone_crs}
\alias{calculate_utm_zone_crs.sf}
\alias{calculate_utm_zone_crs.numeric}
\title{Calculate which utm zones are appropreate for input data}
\usage{
calculate_utm_zone_crs(lat, lon = NULL)

\method{calculate_utm_zone_crs}{sf}(lat, lon = NULL)

\method{calculate_utm_zone_crs}{numeric}(lat, lon = NULL)
}
\arguments{
\item{lat}{vector of latitudes or an sf object}

\item{lon}{vector of longitudes (empty in case x is an sf object)}
}
\value{
vector of utm crs of the same length as input indicating
which zone the data fall into
}
\description{
Calculate which utm zones are appropreate for input data
}
\examples{
trackframe::calculate_utm_zone_crs(
  lat = c(47.6839952, 47.6839941, 47.6839939, 38.5382329, 38.5382306),
  lon = c(9.175119, 9.17498, 9.175254, -121.764227, -121.764351)
)

trackframe::calculate_utm_zone_crs(
  trackframe::move2_mini
)
}
