% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{cbind.trackframe}
\alias{cbind.trackframe}
\title{Combine trackframes by columns}
\usage{
\method{cbind}{trackframe}(...)
}
\arguments{
\item{...}{objects of class trackframe. Other R objects may be coerced as appropriate.}
}
\value{
an object of class trackframe
}
\description{
Take a sequence of trackframes, data.frame, data.table, tibble to combine columns.
}
\examples{
tf1 <- trackframe::tf_mini
tf2 <- tf1
tf2$id2 <- "A"
tf1_tf2 <- cbind(tf1, tf2)
tf1_tf2
class(tf1_tf2)
}
