% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_trackframe}
\alias{check_trackframe}
\title{Checkmate style check if x is a trackframe object}
\usage{
check_trackframe(x, ..., unsorted.ok = TRUE)
}
\arguments{
\item{x}{Object to check}

\item{...}{arguments passed to checkmate::check_data_frame}

\item{unsorted.ok}{Is a trackframe not in order([id], timestamp) allowed? Default: TRUE}
}
\description{
Checkmate style check if x is a trackframe object
}
