% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_frame.R
\name{cocomo_as_tf}
\alias{cocomo_as_tf}
\title{Converts a cocomo object to a track_frame}
\usage{
cocomo_as_tf(x, y, t, ids)
}
\arguments{
\item{x}{matrix of x coordinates (UTM eastings) of all individuals in a group or population (rows) at every time point (columns) x[i,t] gives the x / easting position of individual i at time point t}

\item{y}{matrix of y coordinates (UTM northings) of all individuals in a group or population (rows) at every time point (columns) y[i,t] gives the y / northing position of individual i at time point t}

\item{t}{vector of timestamps in posixct corresponding to the columns of x and y matrices. Timestamps must be uniformly sampled, though it is possible to have gaps (e.g. between different days of recording)}

\item{ids}{data frame giving information about the tracked individuals, with rows correpsonding to the rows of the x and y matrices. There must be one column called id_code which contains a unique individual identifier for each animal (e.g. for meerkats: 'VCVM001', for hyenas: 'WRTH', for coatis: 'Luna') The other columns contained are flexible, and can include information on age, sex, dominance, etc}
}
\value{
an object of class track_frame
}
\description{
Converts a cocomo object to a track_frame
}
\examples{
x <- tf_as_cocomo(sim_travel_paths(3, 3))
cocomo_as_tf(x$x, x$y, x$t, x$ids)
}
