% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{col_guessing}
\alias{col_guessing}
\title{Guesses columns}
\usage{
col_guessing(
  col_names,
  time_col_candidates = tf_options("time_col"),
  easting_col_candidates = tf_options("easting_col"),
  northing_col_candidates = tf_options("northing_col"),
  id_col_candidates = tf_options("id_col")
)
}
\arguments{
\item{col_names}{vector of column names of the input data}

\item{time_col_candidates}{a vector of candidates for time column. Typically provided in tf_options("time_col").}

\item{easting_col_candidates}{a vector of candidates for easting column. Typically provided in tf_options("easting_col").}

\item{northing_col_candidates}{a vector of candidates for northing column. Typically provided in tf_options("northing_col").}

\item{id_col_candidates}{a vector of candidates for id column. Typically provided in tf_options("id_col").}
}
\value{
a list of guesses
}
\description{
Guesses columns
}
