% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{crs}
\alias{crs}
\alias{crs<-}
\title{Extract or Assign from a Track Frame Coordinate Reference System}
\usage{
crs(tf)

crs(tf) <- value
}
\arguments{
\item{tf}{A \code{trackframe} object containing the tracking data.}

\item{value}{a valid projected crs or NA}
}
\value{
A representation of crs extracted from the \code{trackframe}.
 If crs_type is geographic or projected, then crs is a valid input to sf::st_crs.
 If crs_type is nongeoreferenced, then crs can be any non-null value including NA.

tf with new crs
}
\description{
This function retrieves the crs value from a \code{trackframe} object.
}
\examples{
crs(tf_mini)
crs(tf_mini) <- NA
crs(tf_mini)

}
