% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{crs_type}
\alias{crs_type}
\title{Extract Coordinate Reference System Type from a Track Frame}
\usage{
crs_type(tf)
}
\arguments{
\item{tf}{A \code{trackframe} object containing the tracking data.}
}
\value{
One of:
\code{projected} e.g. utm
\code{nongeoreferenced} designed for use in captive or simulated scenarios
}
\description{
This function retrieves the crs value from a \code{trackframe} object.
}
\examples{
crs_type(tf_mini)

}
