% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{deduping}
\alias{deduping}
\title{Deduping of trackframe}
\usage{
deduping(tf, cols = tf_colnames(tf))
}
\arguments{
\item{tf}{an object of class trackframe}

\item{cols}{determines the unique identifier. Default are the key columns provided by
tf_colnames(tf).}
}
\description{
This function removes duplicated entries form objects of class trackframe.
}
\examples{
tf1 <- trackframe::tf_mini
tf2 <- tf1
tf2$northing <- 1:11

tf <- trackframe:::rbind.trackframe(tf1, tf1)
deduping(tf)

tf <- trackframe:::rbind.trackframe(tf1, tf2)
deduping(tf)
deduping(tf, cols = c("time", "id"))
}
