% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{df_demo}
\alias{df_demo}
\title{Artificial Travel Path}
\format{
An object of class \code{data.frame} with 21 rows and 4 columns.
}
\usage{
df_demo
}
\description{
data.frame with manually created data.
}
\keyword{datasets}
