% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{df_mini}
\alias{df_mini}
\title{Simulated Travel Path}
\format{
An object of class \code{data.frame} with 5 rows and 4 columns.
}
\usage{
df_mini
}
\description{
data.frame created with travelpaths::sim_travel_path().
}
\keyword{datasets}
