% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{easting}
\alias{easting}
\title{Extract Easting from a Track Frame}
\usage{
easting(tf)
}
\arguments{
\item{tf}{A `track_frame` object containing the tracking data.
Must have an attribute indicating the easting column (`easting`).}
}
\value{
A vector of easting values extracted from the `track_frame`.
}
\description{
This function retrieves the easting values from a `track_frame` object.
The `track_frame` has a data frame-like structure with an attribute
specifying the column containing easting values.
}
\examples{
tf <- sim_travel_path(10, format = "track_frame")
easting(tf)

}
