% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{get_direction_points}
\alias{get_direction_points}
\title{Obtain direction points}
\usage{
get_direction_points(tf)
}
\arguments{
\item{tf}{an object of class trackframe}
}
\value{
a list with elements x1, y1 providing the x and y coordinates of the direction points
of all different tracks.
Each element of the list is of length of the number of unique id's in the trackframe.
}
\description{
This function obtains direction points (representing the second data point) for all tracks of
objects of class trackframe.
}
