% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{get_starting_points}
\alias{get_starting_points}
\title{Obtain starting points}
\usage{
get_starting_points(tf)
}
\arguments{
\item{tf}{an object of class trackframe}
}
\value{
a trackframe providing the x and y coordinates of the starting points all
different tracks sorted by id (if available) and time.
}
\description{
This function obtains starting points for all tracks in a trackframe object.
}
\examples{
library(trackframe)
get_starting_points(tf_mini)

}
