% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{move2_mini}
\alias{move2_mini}
\title{Simulated Travel Path}
\format{
An object of class \code{move2} (inherits from \code{sf}, \code{data.frame}) with 5 rows and 3 columns.
}
\usage{
move2_mini
}
\description{
move2 object created with travelpaths::sim_travel_path().
}
\keyword{datasets}
