% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{path_data_frame}
\alias{path_data_frame}
\title{Simulated Travel Path}
\format{
An object of class \code{data.frame} with 1000 rows and 4 columns.
}
\usage{
path_data_frame
}
\description{
data.frame created with travelpaths::sim_travel_path().
}
\keyword{datasets}
