% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{path_move2}
\alias{path_move2}
\title{Simulated Travel Path}
\format{
An object of class \code{move2} (inherits from \code{sf}, \code{data.frame}) with 1000 rows and 3 columns.
}
\usage{
path_move2
}
\description{
move2 object created with travelpaths::sim_travel_path().
}
\keyword{datasets}
