% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{path_trackframe}
\alias{path_trackframe}
\title{Simulated Travel Path}
\format{
An object of class \code{trackframe} (inherits from \code{data.frame}) with 1000 rows and 4 columns.
}
\usage{
path_trackframe
}
\description{
trackframe created with travelpaths::sim_travel_path().
}
\keyword{datasets}
