% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{paths_matrix}
\alias{paths_matrix}
\title{Simulated Travel Path with multiple IDs}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3000 rows and 4 columns.
}
\usage{
paths_matrix
}
\description{
matrix created with travelpaths::sim_travel_paths().
}
\keyword{datasets}
