% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{paths_move2}
\alias{paths_move2}
\title{Simulated Travel Path with multiple IDs}
\format{
An object of class \code{move2} (inherits from \code{sf}, \code{data.frame}) with 3000 rows and 3 columns.
}
\usage{
paths_move2
}
\description{
move2 object created with travelpaths::sim_travel_paths().
}
\keyword{datasets}
