% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{paths_sftrack}
\alias{paths_sftrack}
\title{Simulated Travel Path with multiple IDs}
\format{
An object of class \code{sftrack} (inherits from \code{sf}, \code{data.frame}) with 3000 rows and 4 columns.
}
\usage{
paths_sftrack
}
\description{
sftrack object created with travelpaths::sim_travel_paths().
}
\keyword{datasets}
