% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.trackframe}
\alias{plot.trackframe}
\title{Plot trackframes}
\usage{
\method{plot}{trackframe}(
  x,
  direction = FALSE,
  arrow_length = 0.1,
  arrow_code = 2,
  arrow_col = "black",
  arrow_lty = 3,
  arrow_lwd = 1,
  nfacet_col = NULL,
  start_point = FALSE,
  start_point_col = "green",
  start_point_pch = 0,
  start_point_cex = 1,
  end_point = FALSE,
  end_point_col = "red",
  end_point_pch = 1,
  end_point_cex = 1,
  change_point_id = NULL,
  change_point_col = "blue",
  change_point_pch = 4,
  change_point_cex = 1,
  ...
)
}
\arguments{
\item{x}{an object of class \code{trackframe}}

\item{direction}{logical indicator if the path direction should be added to the plot}

\item{arrow_length}{length of the arrow of the direction (argument passed to
\code{\link[graphics]{arrows}})}

\item{arrow_code}{code of the arrow of the direction (argument passed to
\code{\link[graphics]{arrows}})}

\item{arrow_col}{color of the arrow of the direction (argument passed to
\code{\link[graphics]{arrows}})}

\item{arrow_lty}{line type of the arrow of the direction (argument passed to
\code{\link[graphics]{arrows}})}

\item{arrow_lwd}{line width of the arrow of the direction (argument passed to
\code{\link[graphics]{arrows}})}

\item{nfacet_col}{number of columns used in facet.args argument ncol}

\item{start_point}{logical if starting point shoul be plotted}

\item{start_point_col}{color of starting point}

\item{start_point_pch}{pch of starting point}

\item{start_point_cex}{cex of starting point}

\item{end_point}{logical if starting point shoul be plotted}

\item{end_point_col}{color of end point}

\item{end_point_pch}{pch of end point}

\item{end_point_cex}{cex of end point}

\item{change_point_id}{column id of change points if available}

\item{change_point_col}{color of change points}

\item{change_point_pch}{pch of change points}

\item{change_point_cex}{cex of change points}

\item{...}{other arguments used in \code{\link[tinyplot]{tinyplot}}}
}
\description{
Plots coordinates of objects of class \code{\link[trackframe]{trackframe}} based on
\code{\link[tinyplot]{tinyplot}} functionality.
}
\examples{
library(trackframe)
library(tinyplot)

data("tf_mini", package = "trackframe")

data <- tf_mini
class(data)

plot(data)
# set different theme
tinytheme("clean2")
plot(data)

plot(data, direction = TRUE)

track_1 <- select_id(data, "track_1")
plot(track_1)
plot(track_1, direction = TRUE)

}
