% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_coords_by_time}
\alias{plot_coords_by_time}
\title{Plot Time Path}
\usage{
plot_coords_by_time(
  x,
  marker = NULL,
  marker_style = list(col = "blue", pch = 4, cex = 1),
  nfacet_col = NULL,
  mfrow = c(2, 1),
  mar = c(2, 2, 2, 1),
  ...
)
}
\arguments{
\item{x}{an object of class trackframe}

\item{marker}{(optional) column name of additional markers to be plotted (depending on a 0/1
labeling)}

\item{marker_style}{a list where col, pch and cex for markers are specified}

\item{nfacet_col}{number of columns used in facet.args argument ncol}

\item{mfrow}{number of rows used in par()}

\item{mar}{margins used in par()}

\item{...}{other arguments passed to the tinyplot call}
}
\description{
This functions plots x-coordinates as well as y-coordinates over time in separate plots.
}
\examples{
library(trackframe)
plot_coords_by_time(path_trackframe)

path_trackframe$change_points <- 0
path_trackframe$change_points[c(5 , 112, 205, 400, 700)] <- 1
plot_coords_by_time(path_trackframe, marker = "change_points")

}
