% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_time_path}
\alias{plot_time_path}
\title{Plot Time Path}
\usage{
plot_time_path(
  x,
  change_point_id = NULL,
  change_point_col = "blue",
  change_point_pch = 4,
  change_point_cex = 1,
  nfacet_col = NULL,
  mfrow = c(2, 1),
  mar = c(2, 2, 2, 1),
  ...
)
}
\arguments{
\item{x}{an object of class trackframe}

\item{change_point_id}{column id of change points if available}

\item{change_point_col}{color of change points}

\item{change_point_pch}{pch of change points}

\item{change_point_cex}{cex of change points}

\item{nfacet_col}{number of columns used in facet.args argument ncol}

\item{mfrow}{number of rows used in par()}

\item{mar}{margins used in par()}

\item{...}{...}
}
\description{
Plot Time Path
}
