% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{rbind.trackframe}
\alias{rbind.trackframe}
\title{Combine trackframes by rows}
\usage{
\method{rbind}{trackframe}(...)
}
\arguments{
\item{...}{objects of class trackframe.}
}
\value{
an object of class trackframe
}
\description{
Take a sequence of trackframes to combine rows.
}
\examples{
tf1 <- trackframe::tf_mini
tf2 <- tf1
tf2$id <- c(rep("A",5), rep("B", 4), rep("C",2))
tf2$northing <- 20
tf2$easting <- 10
tf1_tf2 <- rbind(tf1, tf2)
tf1_tf2
class(tf1_tf2)
}
