% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{select_id}
\alias{select_id}
\title{Select Tracks by ID from a Track Frame}
\usage{
select_id(tf, id)
}
\arguments{
\item{tf}{A \code{trackframe} object containing the tracking data.
Must have an attribute indicating the track ID column.}

\item{id}{A character or vector of characters representing the track ID(s) to select.}
}
\value{
A filtered \code{trackframe} containing only the specified track(s).
}
\description{
This function filters a trackframe object to include only tracks with the specified ID(s).
It supports selecting a single ID or multiple IDs simultaneously.
}
\examples{
tf <- travelpaths::sim_travel_paths(3, 2:4)
single_track <- select_id(tf, "track_1")
single_track
multiple_tracks <- select_id(tf, c("track_2", "track_3"))
multiple_tracks
}
