% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{set_facet_ncol}
\alias{set_facet_ncol}
\title{Set cols of facet}
\usage{
set_facet_ncol(n)
}
\arguments{
\item{n}{an integer value (corresponding to the number of different ids)}
}
\value{
numeric
}
\description{
Calculates the desired number of columns for trackframe plots using tinyplot engine.
}
\examples{
set_facet_ncol(12)

}
