% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{sftrack_mini}
\alias{sftrack_mini}
\title{Simulated Travel Path}
\format{
An object of class \code{sftrack} (inherits from \code{sf}, \code{data.frame}) with 5 rows and 4 columns.
}
\usage{
sftrack_mini
}
\description{
sftrack object created with travelpaths::sim_travel_path().
}
\keyword{datasets}
