% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{split_by_id}
\alias{split_by_id}
\title{Split trackframe by ID}
\usage{
split_by_id(tf)
}
\arguments{
\item{tf}{A `trackframe` object containing the tracking data.
Must have an attribute indicating the track ID column.}
}
\value{
an object of class `list_of_trackframes` containing a `trackframe` in each list element.
}
\description{
This function splits a trackframe into a list of trackframes by the id.
}
\examples{
tf <- travelpaths::sim_travel_paths(3, 3)
tf_split <- split_by_id(tf)
class(tf_split)
}
