% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_suggestion.R
\name{suggest_utm_zone_crs}
\alias{suggest_utm_zone_crs}
\alias{suggest_utm_zone_crs.sf}
\alias{suggest_utm_zone_crs.numeric}
\title{Suggest a utm crs}
\usage{
suggest_utm_zone_crs(lat, lon = NULL)

\method{suggest_utm_zone_crs}{sf}(lat, lon = NULL)

\method{suggest_utm_zone_crs}{numeric}(lat, lon = NULL)
}
\arguments{
\item{lat}{vector of latitudes or an sf object}

\item{lon}{vector of longitudes (empty in case x is an sf object)}
}
\value{
crs corresponding to the utm zone that the most data points fall into
}
\description{
Uses [calculate_utm_zone_crs] to determine the UTM zone
for each datapoint. Returns the most common zone in the dataset.
}
\details{
No weighting or averaging is done.
Simply the zone that the most of points fall into.
Arbitrary in case of a tie. Future versions may use
a different (better) methodology to chose a zone
when points fall into multiple zones.
}
\examples{
suggest_utm_zone_crs(
  lat = c(47.6839952, 47.6839941, 47.6839939, 38.5382329, 38.5382306),
  lon = c(9.175119, 9.17498, 9.175254, -121.764227, -121.764351)
)

suggest_utm_zone_crs(
  trackframe::move2_mini
)

}
