% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtransform.R
\name{tf_as_sf}
\alias{tf_as_sf}
\alias{tf_as_sftrack}
\alias{tf_as_move2}
\title{Convert a Track Frame to Simple Features (sf) Object}
\usage{
tf_as_sf(tf, ...)

tf_as_sftrack(tf)

tf_as_move2(tf)
}
\arguments{
\item{tf}{A `trackframe` object containing the tracking data. Must have
attributes specifying the easting and northing columns (`easting` and `northing`).}

\item{...}{Additional arguments to be passed to `st_as_sf`.}
}
\value{
An sf object representing the spatial data contained in the `trackframe`.
}
\description{
This function converts a `trackframe` object into a Simple Features (sf) object,
enabling spatial analysis and visualization. The easting and northing columns
are used as coordinates for the sf object.
}
\examples{
sf_object <- tf_as_sf(tf_mini)
print(sf_object)
}
