% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtransform.R
\name{tf_as_xyt}
\alias{tf_as_xyt}
\title{Convert a Track Frame to XYT Format}
\usage{
tf_as_xyt(x, ...)
}
\arguments{
\item{x}{A \code{trackframe} object containing the tracking data.}

\item{...}{other arguments passed to coredata}
}
\value{
A data frame with the easting, northing, time index, and optionally track ID columns.
}
\description{
This function extracts the core spatial-temporal data from a trackframe object,
returning a simplified data frame with just the easting (x), northing (y),
time (t), and optionally the track ID columns.
}
\examples{
tf_as_xyt(tf_mini)
}
