% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtransform.R
\name{tf_backtransform}
\alias{tf_backtransform}
\title{Backtransform}
\usage{
tf_backtransform(tf, all = FALSE)
}
\arguments{
\item{tf}{an object of class \code{trackframe}}

\item{all}{undo all as.trackframe calls, not just the most recent}
}
\value{
an object which has been coerced to \code{trackframe}
}
\description{
Undoes as.trackframe, returning the dataframe to its previous class.
Also attempts to undo sorting performed by as.trackframe.
}
\examples{
library(trackframe)
# for move2
tf <- as.trackframe(data = move2_mini)
tfb <- tf_backtransform(tf)
tfb
move2_mini

# for sftrack
tf <- as.trackframe(data = sftrack_mini)
tfb <- tf_backtransform(tf)
tfb
sftrack_mini

# for data.frame
tf <- as.trackframe(data = df_mini, crs = NA)
tfb <- tf_backtransform(tf)
tfb
df_mini
}
