% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R, R/tf_colnames.R
\name{time_col}
\alias{time_col}
\alias{time_col<-}
\alias{id_col}
\alias{id_col<-}
\alias{easting_col}
\alias{easting_col<-}
\alias{northing_col}
\alias{northing_col<-}
\alias{tf_colnames}
\alias{tf_colnames<-}
\title{Get and set column names of Trackframe Key columns}
\usage{
time_col(tf)

time_col(tf) <- value

id_col(tf)

id_col(tf) <- value

easting_col(tf)

easting_col(tf) <- value

northing_col(tf)

northing_col(tf) <- value

tf_colnames(tf)

tf_colnames(tf) <- value
}
\arguments{
\item{tf}{a trackframe}

\item{value}{new column names. \code{tf_colnames} takes a
named character vector with names \code{easting}, \code{northing}, \code{time}, and (opt) \code{id}.
\code{X_col()} functions take a single string}
}
\value{
X_col returns a character object representing the column name.
tf_colnames returns a named character vector of length 4 indicating the
column names of each of the key columns.
}
\description{
Key columns meaning: easting, northing, time, and (opt) id column
All 3-4 key column names can be accessed or set with \code{tf_colnames()}.
Each also has a dedicated functions to access or set the names of the key columns.
Setting column names in this way will rename the currently configured key column
in the trackframe. It does not change the values of the column.
}
\details{
To identify a different (existing) column as a key column, use \code{as.trackframe}.
}
\examples{
time_col(tf_mini)
time_col(tf_mini) <- as.POSIXct(seq_along(nrow(tf_mini)))
colnames(tf_mini)
id_col(tf_mini)
id_col(tf_mini) <- "track_id"
colnames(tf_mini)
easting_col(tf_mini)
easting_col(tf_mini) <- "x"
northing_col(tf_mini)
northing_col(tf_mini) <- "y"
tf_colnames(tf_mini)
tf_colnames(tf_mini)["id"] <- "animal_id"
}
\seealso{
tf_coords, as.trackframe, tf_id, tf_time
}
