% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{easting}
\alias{easting}
\alias{easting<-}
\alias{northing}
\alias{northing<-}
\title{Extract or Assign Trackframe coordinates}
\usage{
easting(tf)

easting(tf) <- value

northing(tf)

northing(tf) <- value
}
\arguments{
\item{tf}{A \code{trackframe} object containing the tracking data.
Must have an attribute indicating the easting column (`easting`).}

\item{value}{numeric vector of new coordinate values}
}
\value{
A vector of easting or northing values extracted from the \code{trackframe}.
}
\description{
This function retrieves the easting values from a \code{trackframe} object.
The \code{trackframe} has a data frame-like structure with an attribute
specifying the column containing easting values.
}
\examples{
easting(tf_mini)

easting(tf_mini) <- -easting(tf_mini)
northing(tf_mini)

northing(tf_mini) <- northing(tf_mini)*5

}
