% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{id}
\alias{id}
\alias{id<-}
\title{Extract or Assign Trackframe Track ID}
\usage{
id(tf)

id(tf) <- value
}
\arguments{
\item{tf}{A \code{trackframe} object containing the tracking data.
Must have an attribute indicating the track ID column (`id`).}

\item{value}{new values for id column}
}
\value{
A vector of track ID values extracted from the \code{trackframe}.
        NULL if no id column is configured.
}
\description{
This function retrieves the track ID values from a \code{trackframe} object.
The \code{trackframe} has a data frame-like structure with an attribute
specifying the column containing track ID values.
}
\examples{
id(tf_mini)
id(tf_mini) <- rep("new_trackname", nrow(tf_mini))
}
