% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{tf_mini}
\alias{tf_mini}
\title{Simulated Travel Path with multiple IDs}
\format{
An object of class \code{trackframe} (inherits from \code{data.frame}) with 11 rows and 4 columns.
}
\usage{
tf_mini
}
\description{
trackframe object created with travelpaths::sim_travel_paths().
}
\keyword{datasets}
