% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{time.trackframe}
\alias{time.trackframe}
\alias{time<-}
\alias{time<-.trackframe}
\title{Extract or Assign Trackframe Time Index}
\usage{
\method{time}{trackframe}(x, ...)

time(x) <- value

\method{time}{trackframe}(x) <- value
}
\arguments{
\item{x}{A \code{trackframe} object containing the tracking data.
Must have an attribute indicating the time column (`time`).}

\item{...}{...}

\item{value}{vector of POSIX timestamps with length nrow(x)}
}
\value{
A vector of time index values extracted from the \code{trackframe}.
}
\description{
This function retrieves or updates
the time index values from a \code{trackframe} object.
The \code{trackframe} has a data frame-like structure with an attribute
specifying the column containing time index values.
}
\examples{
time(tf_mini)
time(tf_mini) <- seq_along(nrow(tf_mini))

}
