% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{tf_to_sf}
\alias{tf_to_sf}
\title{Convert a Track Frame to Simple Features (sf) Object}
\usage{
tf_to_sf(tf, crs = 4326, ...)
}
\arguments{
\item{tf}{A `track_frame` object containing the tracking data. Must have
attributes specifying the easting and northing columns (`easting` and `northing`).}

\item{crs}{The coordinate reference system (CRS) to be used for the sf object.
Defaults to EPSG code 4326 (WGS84).}

\item{...}{Additional arguments to be passed to `st_as_sf`.}
}
\value{
An sf object representing the spatial data contained in the `track_frame`.
}
\description{
This function converts a `track_frame` object into a Simple Features (sf) object,
enabling spatial analysis and visualization. The easting and northing columns
are used as coordinates for the sf object.
}
\examples{
tf <- sim_travel_path(100, format = "track_frame")
sf_object <- tf_to_sf(tf, crs = 4326)
plot(sf_object)
}
