% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{time}
\alias{time}
\title{Extract Index from a Track Frame}
\usage{
time(tf)
}
\arguments{
\item{tf}{A `track_frame` object containing the tracking data.
Must have an attribute indicating the time column (`time`).}
}
\value{
A vector of time index values extracted from the `track_frame`.
}
\description{
This function retrieves the time index values from a `track_frame` object.
The `track_frame` has a data frame-like structure with an attribute
specifying the column containing time index values.
}
\examples{
tf <- sim_travel_path(100, format = "track_frame")
time(tf)

}
