% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_frame.R
\name{track_frame}
\alias{track_frame}
\title{Create a new Track Frame}
\usage{
track_frame(index, lon, lat, id, ...)
}
\arguments{
\item{index}{Column name of the index column}

\item{data}{A data.frame or tibble or data.table.}

\item{lon_col}{Column name of the longitude column}

\item{lat_col}{Column name of the latitude column}

\item{alt_col}{Column name of the altitude column}

\item{id_col}{Column name of the id column}
}
\value{
A track_frame object
}
\description{
Create a new Track Frame
}
