% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{track_id}
\alias{track_id}
\title{Extract Track ID from a Track Frame}
\usage{
track_id(tf)
}
\arguments{
\item{tf}{A `track_frame` object containing the tracking data.
Must have an attribute indicating the track ID column (`track_id`).}
}
\value{
A vector of track ID values extracted from the `track_frame`.
}
\description{
This function retrieves the track ID values from a `track_frame` object.
The `track_frame` should be a data frame-like structure with an attribute
specifying the column containing track ID values.
}
\examples{
tf <- sim_travel_paths(3, c(2, 4, 5))
track_id_values <- track_id(tf)
print(track_id_values)
}
