% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{type_arrows}
\alias{type_arrows}
\title{Add arrows to a plot}
\usage{
type_arrows(x0, y0, x1, y1, length = 0.25, angle = 30, code = 2)
}
\arguments{
\item{x0}{x0 in \code{\link[graphics]{arrows}}}

\item{y0}{y0 in \code{\link[graphics]{arrows}}}

\item{x1}{x1 in \code{\link[graphics]{arrows}}}

\item{y1}{y1 in \code{\link[graphics]{arrows}}}

\item{length}{length in \code{\link[graphics]{arrows}}}

\item{angle}{angle in \code{\link[graphics]{arrows}}}

\item{code}{code in \code{\link[graphics]{arrows}}}
}
\description{
This function adds arrows to the current (tinyplot) plot. Defines a type for tinyplots.
}
\examples{
library(trackframe)
library(tinyplot)

df_mini
tinyplot(x = df_mini$easting, y = df_mini$northing, type = "l")
tinyplot_add(
  type = type_arrows(
    x0 = df_mini$easting[1],
    y0 = df_mini$northing[1],
    x1 = df_mini$easting[2],
    y1 = df_mini$northing[2],
    length = 0.5,
    code = 2
    ),
  col = "blue",
  lty = 2,
  lwd = 2
)

}
