% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{unique_ids}
\alias{unique_ids}
\title{Extract Unique IDs from a Track Frame}
\usage{
unique_ids(tf)
}
\arguments{
\item{tf}{A \code{trackframe} object containing the tracking data.
Must have an attribute indicating the track ID column (\code{id}).}
}
\value{
A vector of unique track IDs extracted from the \code{trackframe}.
}
\description{
This function retrieves the unique track IDs from a \code{trackframe} object.
The \code{trackframe} should be a data frame-like structure with an attribute
specifying the column containing track IDs.
}
\examples{
unique_ids(tf_mini)

}
\seealso{
\code{\link[=as.trackframe]{as.trackframe()}}
}
