% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackframe.R
\name{trackframe}
\alias{trackframe}
\alias{as.trackframe}
\alias{as.trackframe.data.frame}
\alias{as.trackframe.matrix}
\alias{as.trackframe.move2}
\alias{as.trackframe.sftrack}
\alias{as.trackframe.trackframe}
\title{Convert an object to a \code{track\_frame}}
\usage{
trackframe(
  data,
  time_col = tf_options("time_col"),
  easting_col = tf_options("easting_col"),
  northing_col = tf_options("northing_col"),
  id_col = tf_options("id_col"),
  crs_input = NULL,
  utm_epsg = NULL,
  sort = TRUE,
  coerce_to = "base",
  ...
)

as.trackframe(
  data,
  time_col = tf_options("time_col"),
  easting_col = tf_options("easting_col"),
  northing_col = tf_options("northing_col"),
  id_col = tf_options("id_col"),
  sort = TRUE,
  coerce_to = "base",
  ...
)

\method{as.trackframe}{data.frame}(
  data,
  time_col = tf_options("time_col"),
  easting_col = tf_options("easting_col"),
  northing_col = tf_options("northing_col"),
  id_col = tf_options("id_col"),
  sort = TRUE,
  coerce_to = "base",
  crs_input = NULL,
  utm_epsg = NULL,
  ...
)

\method{as.trackframe}{matrix}(
  data,
  time_col = tf_options("time_col"),
  easting_col = tf_options("easting_col"),
  northing_col = tf_options("northing_col"),
  id_col = tf_options("id_col"),
  sort = TRUE,
  coerce_to = "base",
  crs_input = NULL,
  utm_epsg = NULL,
  ...
)

\method{as.trackframe}{move2}(
  data,
  time_col = NULL,
  easting_col = NULL,
  northing_col = NULL,
  id_col = NULL,
  sort = TRUE,
  coerce_to = "base",
  ...
)

\method{as.trackframe}{sftrack}(
  data,
  time_col = NULL,
  easting_col = NULL,
  northing_col = NULL,
  id_col = NULL,
  sort = TRUE,
  coerce_to = "base",
  ...
)

\method{as.trackframe}{trackframe}(
  data,
  time_col = NULL,
  easting_col = NULL,
  northing_col = NULL,
  id_col = NULL,
  sort = TRUE,
  coerce_to = "base",
  ...
)
}
\arguments{
\item{data}{a `data.frame`, a `matrix`, an `sftrack` object or a `move2` object containing the tracking data.}

\item{time_col}{a character string specifying the column name of the time column. If no column is specified, 
the `time_col` is tried to be matched by possible names provided in `tf_options("time_col")`. In case of multiple matches, the first match is chosen.}

\item{easting_col}{a character string specifying the column name of the easting column. If no column is specified, 
the `easting_col` is tried to be matched by possible names provided in `tf_options("easting_col")`. In case of multiple matches, the first match is chosen.}

\item{northing_col}{a character string specifying the column name of the northing column. If no column is specified, 
the `northing_col` is tried to be matched by possible names provided in `tf_options("northing_col")`. In case of multiple matches, the first match is chosen.}

\item{id_col}{optional character vector specifying identifier column names. If no column is specified, 
the `id_col` is tried to be matched by possible names provided in `tf_options("id_col")`. In case of multiple matches, the first match is chosen.}

\item{crs_input}{crs code for input of coordinates}

\item{utm_epsg}{crs value for utm zone of the `trackframe` output}

\item{sort}{logical, if data should be sorted according to id_col and time_col}

\item{coerce_to}{format trackframe is coerced to. `base`, `data.table` and `tibble` are supported. Deafault is `base` and gives a `data.frame`.}

\item{...}{Additional arguments (unused).}
}
\value{
A `trackframe` object with appropriate attributes set.
}
\description{
This function converts an object into a `trackframe` object,
ensuring required columns exist and have valid data types. Coordinates must be provided in easting/northing.
Coordinates for  `sftrack` and `move2` objects are transformed to easting/northing if possible.
}
\examples{
df <- data.frame(
  time_col = as.POSIXct(Sys.time() + 1:5),
  easting_col = runif(5, 0, 10),
  northing_col = runif(5, 0, 10),
  id = 1:5
)
tf <- trackframe(df, time_col = "time_col", easting_col = "easting_col",
                        northing_col = "northing_col", id_col = "id")
attributes(tf)
df <- data.frame(
  time_col = as.POSIXct(Sys.time() + 1:5),
  easting_col = runif(5, 0, 10),
  northing_col = runif(5, 0, 10),
  id = 1:5
)
tf <- as.trackframe(df, time_col = "time_col", easting_col = "easting_col",
                        northing_col = "northing_col", id_col = "id")
attributes(tf)

set.seed(2025)
sim_df <- travelpaths::sim_travel_path(5, format = "data.frame")
as.trackframe(sim_df, crs_input = 4326)

set.seed(2025)
df <- data.frame(
  x = rnorm(10),
    y = rnorm(10),
      t = 1:10,
        animal_id = c(rep('a', 5), rep('b', 5))
        )
 as.trackframe(df,
               time_col = "t",
               easting_col = "x",
               northing_col = "y",
               id_col = "animal_id")
 # with col guessing
 as.trackframe(df, coerce_to = "base")
 
 tf_df <- as.trackframe(df, coerce_to = "base")
 tf_dt <- as.trackframe(df, coerce_to = "data.table")
 tf_tib <- as.trackframe(df, coerce_to = "tibble")
 
 tf_backtransform(tf_df)
 tf_backtransform(tf_dt)
 tf_backtransform(tf_tib)

# example for move2 objects
library(move2)
library(trackframe)
albatross_move2 <- mt_read(mt_example()) |>
  sf::st_transform(3857)
 albatross_move2 <- albatross_move2[!sf::st_is_empty(albatross_move2),]
 albatross_tf <- as.trackframe(albatross_move2)
class(albatross_tf)

# example for sftrack objects
library(sftrack)
library(trackframe)
data("raccoon", package = "sftrack")
raccoon$month <- as.POSIXlt(raccoon$timestamp)$mon + 1
raccoon$time <- as.POSIXct(raccoon$timestamp, tz = "EST")
coords <- c("longitude","latitude")
group <- list(id = raccoon$animal_id, month = as.POSIXlt(raccoon$timestamp)$mon+1)
time <- "time"
error <- "fix"
crs <- 4326
# create a sftrack object
my_sftrack <- as_sftrack(data = raccoon,
                         coords = coords,
                         group = group,
                         time = time,
                         error = error,
                         crs = crs)

sftrack_tf <- as.trackframe(my_sftrack)
class(sftrack_tf)

}
