% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackframe.R
\name{trackframe}
\alias{trackframe}
\alias{as.trackframe}
\alias{as.trackframe.data.frame}
\alias{as.trackframe.matrix}
\alias{as.trackframe.move2}
\alias{as.trackframe.sftrack}
\alias{as.trackframe.trackframe}
\title{Convert an object to a \code{trackframe}}
\usage{
trackframe(
  data,
  time_col = tf_options("time_col"),
  easting_col = tf_options("easting_col"),
  northing_col = tf_options("northing_col"),
  id_col = tf_options("id_col"),
  sort = TRUE,
  coerce_to = "base",
  verbose = TRUE,
  crs_input = NULL,
  utm_epsg = NULL,
  ...
)

as.trackframe(
  data,
  time_col = tf_options("time_col"),
  easting_col = tf_options("easting_col"),
  northing_col = tf_options("northing_col"),
  id_col = tf_options("id_col"),
  sort = TRUE,
  coerce_to = "base",
  verbose = TRUE,
  ...
)

\method{as.trackframe}{data.frame}(
  data,
  time_col = tf_options("time_col"),
  easting_col = tf_options("easting_col"),
  northing_col = tf_options("northing_col"),
  id_col = tf_options("id_col"),
  sort = TRUE,
  coerce_to = "base",
  verbose = TRUE,
  crs_input = NULL,
  utm_epsg = NULL,
  ...
)

\method{as.trackframe}{matrix}(
  data,
  time_col = tf_options("time_col"),
  easting_col = tf_options("easting_col"),
  northing_col = tf_options("northing_col"),
  id_col = tf_options("id_col"),
  sort = TRUE,
  coerce_to = "base",
  verbose = TRUE,
  crs_input = NULL,
  utm_epsg = NULL,
  ...
)

\method{as.trackframe}{move2}(
  data,
  time_col = NULL,
  easting_col = NULL,
  northing_col = NULL,
  id_col = NULL,
  sort = TRUE,
  coerce_to = "base",
  verbose = TRUE,
  ...
)

\method{as.trackframe}{sftrack}(
  data,
  time_col = NULL,
  easting_col = NULL,
  northing_col = NULL,
  id_col = NULL,
  sort = TRUE,
  coerce_to = "base",
  verbose = TRUE,
  ...
)

\method{as.trackframe}{trackframe}(data, ...)
}
\arguments{
\item{data}{a \code{data.frame}, a \code{matrix}, an \code{sftrack} object or a \code{move2} object containing the tracking data.}

\item{time_col}{a character string specifying the column name of the time column. If no column is specified, 
the `time_col` is tried to be matched by possible names provided in `tf_options("time_col")`. In case of multiple matches, the first match is chosen.}

\item{easting_col}{a character string specifying the column name of the easting column. If no column is specified, 
the `easting_col` is tried to be matched by possible names provided in `tf_options("easting_col")`. In case of multiple matches, the first match is chosen.}

\item{northing_col}{a character string specifying the column name of the northing column. If no column is specified, 
the `northing_col` is tried to be matched by possible names provided in `tf_options("northing_col")`. In case of multiple matches, the first match is chosen.}

\item{id_col}{optional character vector specifying identifier column names. If no column is specified, 
the `id_col` is tried to be matched by possible names provided in `tf_options("id_col")`. In case of multiple matches, the first match is chosen.}

\item{sort}{logical, if data should be sorted according to id_col and time_col}

\item{coerce_to}{the format trackframe is coerced to. `base`, `data.table` and `tibble` are supported. Default is `base` and coerces to a `data.frame`.}

\item{verbose}{logical, default value is \code{TRUE}}

\item{crs_input}{crs code for input of coordinates}

\item{utm_epsg}{crs value for utm zone of the \code{trackframe} output}

\item{...}{Additional arguments (unused).}
}
\value{
A \code{trackframe} object with appropriate attributes set.
}
\description{
This function converts an object into a \code{trackframe} object,
ensuring required columns exist and have valid data types. Coordinates must be provided in easting/northing.
Coordinates for  \code{data.frame}, \code{sftrack} and \code{move2} objects are transformed to easting/northing if possible.
}
\examples{
library(trackframe)
df <- data.frame(
  time_col = as.POSIXct(Sys.time() + 1:5),
  easting_col = runif(5, 0, 10),
  northing_col = runif(5, 0, 10),
  id = 1:5
)
tf <- trackframe(df, time_col = "time_col", easting_col = "easting_col",
                        northing_col = "northing_col", id_col = "id")
attributes(tf)
df <- data.frame(
  time_col = as.POSIXct(Sys.time() + 1:5),
  easting_col = runif(5, 0, 10),
  northing_col = runif(5, 0, 10),
  id = 1:5
)
tf <- as.trackframe(df, time_col = "time_col", easting_col = "easting_col",
                        northing_col = "northing_col", id_col = "id")
class(tf)
attributes(tf)
easting(tf)
northing(tf)
time(tf)
id(tf)

set.seed(2025)
sim_df <- travelpaths::sim_travel_path(5, format = "data.frame")
as.trackframe(sim_df, crs_input = 4326)

set.seed(2025)
df <- data.frame(
  x = rnorm(10),
    y = rnorm(10),
      t = 1:10,
        animal_id = c(rep('a', 5), rep('b', 5))
        )
 as.trackframe(df,
               time_col = "t",
               easting_col = "x",
               northing_col = "y",
               id_col = "animal_id")
 # with col guessing
 as.trackframe(df, coerce_to = "base")
 
 tf_df <- as.trackframe(df, coerce_to = "base")
 tf_dt <- as.trackframe(df, coerce_to = "data.table")
 tf_tib <- as.trackframe(df, coerce_to = "tibble")
 
 tf_backtransform(tf_df)
 tf_backtransform(tf_dt)
 tf_backtransform(tf_tib)

# example for move2 objects
library(move2)
library(trackframe)
set.seed(2025)
move2_dat <- travelpaths::sim_travel_path(5, format = "move2")
tf <- as.trackframe(data = move2_dat)

move2_dat2 <- tf_backtransform(tf)
all.equal(move2_dat, move2_dat2)

# example for sftrack objects
library(sftrack)
set.seed(2025)
sftrack_dat <- travelpaths::sim_travel_path(5, format = "sftrack")
tf <- as.trackframe(data = sftrack_dat)
sftrack_dat2 <- tf_backtransform(tf)
sftrack_dat2$id <- unlist(sftrack_dat2$sft_group)
all.equal(sftrack_dat2, sftrack_dat)

}
