% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackframe.R
\name{trackframe}
\alias{trackframe}
\alias{as.trackframe}
\alias{as.trackframe.data.frame}
\alias{as.trackframe.matrix}
\alias{as.trackframe.move2}
\alias{as.trackframe.sftrack}
\alias{as.trackframe.trackframe}
\title{Convert an object to a \code{trackframe}}
\usage{
trackframe(
  data,
  time_col = tf_options("time_col"),
  easting_col = tf_options("easting_col"),
  northing_col = tf_options("northing_col"),
  id_col = tf_options("id_col"),
  sort = TRUE,
  coerce_to = "base",
  crs = NULL,
  ...
)

as.trackframe(
  data,
  time_col = tf_options("time_col"),
  easting_col = tf_options("easting_col"),
  northing_col = tf_options("northing_col"),
  id_col = tf_options("id_col"),
  sort = TRUE,
  coerce_to = "base",
  crs = NULL,
  ...
)

\method{as.trackframe}{data.frame}(
  data,
  time_col = tf_options("time_col"),
  easting_col = tf_options("easting_col"),
  northing_col = tf_options("northing_col"),
  id_col = tf_options("id_col"),
  sort = TRUE,
  coerce_to = "base",
  crs = NULL,
  ...
)

\method{as.trackframe}{matrix}(
  data,
  time_col = tf_options("time_col"),
  easting_col = tf_options("easting_col"),
  northing_col = tf_options("northing_col"),
  id_col = tf_options("id_col"),
  sort = TRUE,
  coerce_to = "base",
  crs = NULL,
  ...
)

\method{as.trackframe}{move2}(
  data,
  time_col = NULL,
  easting_col = NULL,
  northing_col = NULL,
  id_col = NULL,
  sort = TRUE,
  coerce_to = "base",
  ...
)

\method{as.trackframe}{sftrack}(
  data,
  time_col = NULL,
  easting_col = NULL,
  northing_col = NULL,
  id_col = NULL,
  sort = TRUE,
  coerce_to = "base",
  ...
)

\method{as.trackframe}{trackframe}(data, ...)
}
\arguments{
\item{data}{a \code{data.frame}, a \code{matrix}, an \code{sftrack} object or a \code{move2} object
containing the tracking data.}

\item{time_col}{a character string or vector specifying the column name of the time column.}

\item{easting_col}{a character string or vector specifying the column name of the easting column.
If a vector is provided, the first element matching a column name of `data`` is used.}

\item{northing_col}{a character string or vector specifying the column name of
the northing column.
If a vector is provided, the first element matching a column name of `data`` is used.}

\item{id_col}{optional character vector specifying identifier column names.
If no column is specified,
If a vector is provided, the first element matching a column name of `data`` is used.}

\item{sort}{logical, if data should be sorted by id_col and time_col.}

\item{coerce_to}{the type of dataframe that trackframe is coerced to.
`base`, `data.table` and `tibble` are supported.
Default is `base` and coerces to a `data.frame` without `data.table` or `tbl` classes.
If NULL, the returned `trackframe` object takes the same dataframe type as `data` argument}

\item{crs}{coordinate reference system code for of coordinates. Required for non-sf input.
Use `NA` to denote non georeferenced cartesian coordinates.}

\item{...}{Additional arguments (unused).}
}
\value{
A \code{trackframe} object with appropriate attributes set.
}
\description{
This function converts an object into a \code{trackframe} object,
checks that required columns exist and have valid data types.
Coordinates must be provided in easting/northing. Coordinates
coordinate system must be cartesian: either a projected coordinate system
for georeference data or another cartesian coordinate system for non-
georeferenced data (e.g. for captive or simulated data).
}
\examples{
library(trackframe)
df <- data.frame(
  time_col = as.POSIXct(Sys.time() + 1:5),
  easting_col = runif(5, 0, 10),
  northing_col = runif(5, 0, 10),
  id = 1:5
)
tf <- trackframe(df, time_col = "time_col", easting_col = "easting_col",
  northing_col = "northing_col", id_col = "id", crs = NA)

attributes(tf)
df <- data.frame(
  time_col = as.POSIXct(Sys.time() + 1:5),
  easting_col = runif(5, 0, 10),
  northing_col = runif(5, 0, 10),
  id = 1:5
)
tf <- as.trackframe(df, time_col = "time_col", easting_col = "easting_col",
                        northing_col = "northing_col", id_col = "id", crs = NA)
class(tf)
attributes(tf)
easting(tf)
northing(tf)
time(tf)
id(tf)

as.trackframe(df_mini, crs = NA)

set.seed(2025)
df <- data.frame(
  x = rnorm(10),
  y = rnorm(10),
  t = 1:10,
  animal_id = c(rep('a', 5), rep('b', 5))
)

 # crs = NA indicates non georeferenced coordinates
 as.trackframe(
  df,
  time_col = "t",
  easting_col = "x",
  northing_col = "y",
  id_col = "animal_id",
  crs = NA
)
 # with col guessing
 as.trackframe(df, coerce_to = "base", crs = NA)

 tf_df <- as.trackframe(df, coerce_to = "base", crs = NA)
 tf_dt <- as.trackframe(df, coerce_to = "data.table", crs = NA)
 tf_tib <- as.trackframe(df, coerce_to = "tibble", crs = NA)

 tf_backtransform(tf_df)
 tf_backtransform(tf_dt)
 tf_backtransform(tf_tib)

 tf_df <- as.trackframe(df, coerce_to = "base", crs = "EPSG:32632")
 # Non cartesian/lat long crs e.g. "EPSG:4326" is not acceptable.

# example for move2 objects
library(move2)
library(trackframe)

# column mapping and crs are implicit so do not need to be specified in args.
tf <- as.trackframe(data = move2_mini)

move2_dat2 <- tf_backtransform(tf)
all.equal(move2_dat2, move2_mini)

# example for sftrack objects
library(sftrack)
set.seed(2025)
tf <- as.trackframe(data = sftrack_mini)
sftrack_dat2 <- tf_backtransform(tf)
sftrack_dat2$id <- unlist(sftrack_dat2$sft_group)
all.equal(sftrack_dat2, sftrack_mini)

}
