% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocomo_transformations.R
\name{cocomo_as_tf}
\alias{cocomo_as_tf}
\title{Converts a cocomo object to a trackframe}
\usage{
cocomo_as_tf(
  xs,
  ys,
  t,
  ids,
  crs = NA,
  na_omit = TRUE,
  sort = TRUE,
  coerce_to = "base",
  verbose = FALSE
)
}
\arguments{
\item{xs}{matrix of x coordinates (UTM eastings) of all individuals
in a group or population (rows) at every time point (columns)
x[i,t] gives the x / easting position of individual i at time point t}

\item{ys}{matrix of y coordinates (UTM northings) of all individuals
in a group or population (rows) at every time point (columns)
y[i,t] gives the y / northing position of individual i at time point t}

\item{t}{vector of timestamps in posixct corresponding to the columns
of x and y matrices. Timestamps must be uniformly sampled,
though it is possible to have gaps (e.g. between different days of recording)}

\item{ids}{data frame giving information about the tracked individuals,
with rows correpsonding to the rows of the x and y matrices.
There must be one column called id_code which contains a unique
individual identifier for each animal (e.g. for meerkats:
'VCVM001', for hyenas: 'WRTH', for coatis: 'Luna')
The other columns contained are flexible,
and can include information on age, sex, dominance, etc}

\item{crs}{coordinate reference system}

\item{na_omit}{logical indicator if NAs should be omitted}

\item{sort}{logical, if data should be sorted according to id_col and time_col}

\item{coerce_to}{the format trackframe is coerced to. `base`,
`data.table` and `tibble` are supported. Default is `base` and coerces to a `data.frame`.}

\item{verbose}{logical, default value is \code{TRUE}}
}
\value{
an object of class trackframe
}
\description{
Converts a cocomo object to a trackframe
}
\examples{
cocomo <- tf_as_cocomo(tf_mini)
cocomo_as_tf(cocomo$xs, cocomo$ys, cocomo$t, cocomo$ids)
}
