% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_arrow_points}
\alias{get_arrow_points}
\title{Obtain arrow points of starting points}
\usage{
get_arrow_points(tf, sort = TRUE)
}
\arguments{
\item{tf}{an object of class trackframe}

\item{sort}{a logical indicator}
}
\description{
Obtain arrow points of starting points
}
