% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{get_direction_points}
\alias{get_direction_points}
\title{Obtain direction points}
\usage{
get_direction_points(tf)
}
\arguments{
\item{tf}{an object of class trackframe}
}
\value{
a trackframe providing the x and y coordinates of the direction points
of all different tracks sorted by id (if available) and time.
}
\description{
This function obtains direction points (representing the second data point) for all tracks of
objects of class trackframe.
}
\examples{
library(trackframe)
get_direction_points(tf_mini)

}
