% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{get_starting_points}
\alias{get_starting_points}
\title{Obtain starting points}
\usage{
get_starting_points(tf)
}
\arguments{
\item{tf}{an object of class trackframe}
}
\value{
a list with elements x0, y0 providing the x and y coordinates of th starting points all
different tracks.
Each element of the list is of length of the number of unique id's.
}
\description{
This function obtains starting points for all tracks of objects of class trackframe.
}
