% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{merge.trackframe}
\alias{merge.trackframe}
\title{Merge Two trackframes}
\usage{
\method{merge}{trackframe}(x, y, sort = TRUE, suffixes = c("", ".y"), by = NULL, ...)
}
\arguments{
\item{x}{an object of class trackframe}

\item{y}{an object of class trackframe}

\item{sort}{logical, if true (default), data will be sorted by id_col and then by time_col.}

\item{suffixes}{a character vector of length 2 specifying the suffixes to be used for making
unique the names of columns in the result which are not used for merging (appearing in by etc).}

\item{by}{specifications of the columns used for merging. Default is tf_colnames(x)}

\item{...}{other arguments to be passed to appropriate merge methods}
}
\value{
an object of class trackframe
}
\description{
Merge two trackframes by key colums ("time", "id", "easting", "northing"), or do other versions
of database join operations.
}
\examples{
tf1 <- trackframe::tf_mini
tf2 <- tf1
tf2$id2 <- "A"
tf1_tf2 <- merge(tf1, tf2)
tf1_tf2
class(tf1_tf2)

tf3 <- tf1
tf3$id <- c(rep("A",5), rep("B", 4), rep("C",2))
merge(tf1, tf3, all = TRUE)
}
