% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.trackframe}
\alias{plot.trackframe}
\title{Plot trackframes}
\usage{
\method{plot}{trackframe}(
  x,
  direction = FALSE,
  arrow_style = list(length = 0.1, code = 2, col = "black", lty = 3, lwd = 1),
  nfacet_col = NULL,
  start_point = FALSE,
  start_point_style = list(col = "green", pch = 0, cex = 1),
  end_point = FALSE,
  end_point_style = list(col = "red", pch = 1, cex = 1),
  marker = NULL,
  marker_style = list(col = "blue", pch = 4, cex = 1),
  ...
)
}
\arguments{
\item{x}{an object of class \code{trackframe}}

\item{direction}{logical indicator if the path direction should be added to the plot}

\item{arrow_style}{a list of length, code, col, lty, lwd of the arrow of the direction (argument passed to
\code{\link[graphics]{arrows}}) specifying the style of the arrows}

\item{nfacet_col}{number of columns used in facet.args argument ncol}

\item{start_point}{logical if starting point should be plotted}

\item{start_point_style}{a list where col, pch and cex for starting points are specified}

\item{end_point}{logical if starting point should be plotted}

\item{end_point_style}{a list where col, pch and cex for end points are specified}

\item{marker}{(optional) column name of additional markers to be ploted (depending on a 0/1
labeling)}

\item{marker_style}{a list where col, pch and cex for markers are specified}

\item{...}{other arguments used in \code{\link[tinyplot]{tinyplot}}}
}
\description{
Plots coordinates of objects of class \code{\link[trackframe]{trackframe}} based on
\code{\link[tinyplot]{tinyplot}} functionality.
}
\examples{
library(trackframe)

data("tf_mini", package = "trackframe")

data <- tf_mini
class(data)

plot(data)
# set different theme
library(tinyplot)
tinytheme("clean2")
plot(data)

plot(data, direction = TRUE)

track_1 <- select_id(data, "track_1")
plot(track_1)
plot(track_1, direction = TRUE)

plot(track_1, start_point = TRUE, start_point_style = list(col = "blue"), end_point = TRUE)

}
