% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.trackframe}
\alias{plot.trackframe}
\title{Plot trackframes}
\usage{
\method{plot}{trackframe}(
  x,
  lines = TRUE,
  lines_style = list(alpha = 0.5),
  points = TRUE,
  points_style = list(pch = 19),
  start_indicator = TRUE,
  start_indicator_style = list(arrowhead_loc = 0.001, length = 0.1, lwd = 3),
  end_indicator = TRUE,
  end_indicator_style = list(arrowhead_loc = 0.001, length = 0.1, code = 1, angle = 90,
    lwd = 3),
  marker = NULL,
  marker_style = list(pch = 22, col = "black", cex = 2),
  facet = FALSE,
  facet.args = list(),
  theme = NULL,
  xlab = sprintf("easting (\%s)", easting_col(x)),
  ylab = sprintf("northing (\%s)", northing_col(x)),
  asp = 1,
  ...
)
}
\arguments{
\item{x}{an object of class \code{trackframe}}

\item{lines}{logical should lines be plotted?}

\item{lines_style}{list of graphical parameters
passed to \code{\link[tinyplot:tinyplot]{tinyplot::tinyplot()}} when plotting lines}

\item{points}{logical should points be plotted?}

\item{points_style}{graphical parameters
passed to \code{\link[tinyplot:tinyplot]{tinyplot::tinyplot()}} when plotting points}

\item{start_indicator}{logical indicator if arrows indicating
start point of each track should be added}

\item{start_indicator_style}{list of graphical parameters for start indicator.
See section 'x_indicator_style'.}

\item{end_indicator}{logical indicator if arrows indicating
endpoint of each track should be added}

\item{end_indicator_style}{list of graphical parameters for end indicator.
See section 'x_indicator_style'.}

\item{marker}{character column name of logical column indicating which points to mark}

\item{marker_style}{list of graphical parameters passed to \code{\link[tinyplot]{tinyplot}}
as markers are being drawn}

\item{facet}{logical should the plot be facetted by track?}

\item{facet.args}{list of arguments controlling facet behavior.
If \code{ncol} is unspecified, an attempt is made chose a visually pleasing value.
See \code{\link[tinyplot]{tinyplot}}}

\item{theme}{character or list:
\enumerate{
\item \code{NULL} (default): Use currently set \code{tinyplot} theme if one is set.
Otherwise use trackframe default.
\item a string naming a built-in tinyplot theme \code{vignette("themes", package = "tinyplot")} or
\item a list of graphical parameters defining a custom theme
}}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the x axis}

\item{asp}{numeric y/x aspect ratio}

\item{...}{additional graphical parameters passed to \code{\link[tinyplot]{tinyplot}}}
}
\value{
No return value, called for side effect of producing a plot.
}
\description{
Plots coordinates of objects of class \code{\link[trackframe]{trackframe}} based on
\code{\link[tinyplot]{tinyplot}} functionality.
}
\section{x_style}{
The arguments \code{lines_style}, \code{points_style},
\code{start_indicator_style}, \code{end_indicator_style} and \code{marker_style} each take
list of argments passed to \code{tinyplot}to style the corresponding element.

Possible arguments can come from:
\itemize{
\item \code{\link[tinyplot]{tinyplot}} (e.g. \code{facet.args})
\item \code{\link[graphics]{par}} (e.g. \code{fg})
}

Parameters in \code{...} can come from the same lists and are applied to
the entire plot.
}

\section{x_indicator_style}{
Start/end indicators are arrows that point from the start/endpoint of each track
toward the second/penultimate point.
In addition to standard arrow paramenters,
\code{start_indicator_style} and \code{end_indicator_style} can have a key: \code{arrowhead_loc}
which controls how far along the first/last segment the arrow extends.
\code{0} refering to putting the arrow having the head only.
and \code{1} referring to having the arrow extend to second/penultimate point.
.001 inch is added to the user-provided value to ensure arrow direction
can be determined.

Note: Because end arrows go from the endpoint toward the penultimate point,
the \code{\link[graphics:arrows]{graphics::arrows()}} default of \verb{angle = 30, code = 2} will
result in arrows pointing against the direction of travel.
}

\examples{
library(trackframe)

data("tf_mini", package = "trackframe")

data <- tf_mini
class(data)

plot(data)
# set different theme
library(tinyplot)
tinytheme("clean2")
plot(data)

plot(data)


# with facets and no indicator arrows
plot(
  data,
  facet = TRUE,
  start_indicator = FALSE,
  end_indicator = FALSE
)

# allow free y axis
plot(data, facet.args = list("free" = TRUE))

track_1 <- select_id(data, "track_1")
plot(track_1)
plot(track_1, direction = TRUE)

# Customize start/stop arrows
plot(track_1, start_indicator_style = list(col = "blue"))

}
