% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{set_facet_ncol}
\alias{set_facet_ncol}
\title{Set cols of facet}
\usage{
set_facet_ncol(n)
}
\arguments{
\item{n}{an integer value (corresponding to the number of differents ids)}
}
\value{
an integer which determines the number of facet columns used in tinyplot
}
\description{
calculates the number of columns specified for plots using tinyplot engine.
}
