% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{sort.trackframe}
\alias{sort.trackframe}
\title{Sorting of Trackframes}
\usage{
\method{sort}{trackframe}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{an object of class trackframe}

\item{decreasing}{logical indicating if sort should be increasing or decreasing.}

\item{...}{additional arguments passed to order}
}
\value{
an object of class trackframe
}
\description{
Sort a trackframe into ascending or descending order. Sorts if available by track id and time
columns of the trackframe. If no id column is available only sorted by time.
}
