% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{split_by_id}
\alias{split_by_id}
\title{Split trackframe by ID}
\usage{
split_by_id(tf)
}
\arguments{
\item{tf}{A \code{trackframe} object containing the tracking data.
Must have an attribute indicating the track ID column.}
}
\value{
an object of class `list_of_trackframes` containing a \code{trackframe}
in each list element.
}
\description{
This function splits a \code{trackframe} object into a list of trackframes by the id.
}
\examples{
tf_split <- split_by_id(tf_mini)
class(tf_split)
}
