% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_suggestion.R
\name{suggest_utm_crs}
\alias{suggest_utm_crs}
\alias{suggest_utm_crs.sf}
\alias{suggest_utm_crs.numeric}
\title{Suggest a utm crs}
\usage{
suggest_utm_crs(lat, lon = NULL)

\method{suggest_utm_crs}{sf}(lat, lon = NULL)

\method{suggest_utm_crs}{numeric}(lat, lon = NULL)
}
\arguments{
\item{lat}{vector of latitudes or an sf object}

\item{lon}{vector of longitudes (empty in case x is an sf object)}
}
\value{
crs corresponding to the utm zone that the most data points fall into
}
\description{
Suggest a utm crs
}
\details{
arbitrary in case of a tie
}
