% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{tf_as_sf}
\alias{tf_as_sf}
\alias{tf_as_sftrack}
\alias{tf_as_move2}
\title{Convert a Track Frame to Simple Features (sf) Object}
\usage{
tf_as_sf(tf, tf_crs = NULL, crs_new = NULL, ...)

tf_as_sftrack(tf, tf_crs = NULL, crs_new = NULL, ...)

tf_as_move2(tf, tf_crs = NULL, crs_new = NULL, ...)
}
\arguments{
\item{tf}{A `track_frame` object containing the tracking data. Must have
attributes specifying the easting and northing columns (`easting` and `northing`).}

\item{tf_crs}{The coordinate reference system (CRS) used in the `track_frame`.}

\item{crs_new}{The coordinate reference system (CRS) to be used for the sf object.}

\item{...}{Additional arguments to be passed to `st_as_sf`.}
}
\value{
An sf object representing the spatial data contained in the `track_frame`.
}
\description{
This function converts a `track_frame` object into a Simple Features (sf) object,
enabling spatial analysis and visualization. The easting and northing columns
are used as coordinates for the sf object.
}
\examples{
tf <- travelpaths::sim_travel_paths(4, 2:5)
sf_object <- tf_as_sf(tf, tf_crs = 32610, crs_new = 4326)
print(sf_object)
}
