% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtransform.R
\name{tf_backtransform}
\alias{tf_backtransform}
\title{Title}
\usage{
tf_backtransform(tf)
}
\arguments{
\item{tf}{an object of class `trackframe`}
}
\value{
an object which has been coerced to `trackframe`
}
\description{
Title
}
\examples{
library(travelpaths)
library(trackframe)
# for move2
tf <- as.trackframe(data = move2_mini)
tfb <- tf_backtransform(tf)
tfb
move2_mini

# for sftrack
tf <- as.trackframe(data = sftrack_mini)
tfb <- tf_backtransform(tf)
tfb
sftrack_mini

# for data.frame
tf <- as.trackframe(data = df_mini)
tfb <- tf_backtransform(tf)
tfb
df_mini
}
