% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtransform.R
\name{tf_backtransform}
\alias{tf_backtransform}
\title{Title}
\usage{
tf_backtransform(tf)
}
\arguments{
\item{tf}{an object of class `trackframe`}
}
\value{
an object which has been coerced to `trackframe`
}
\description{
Title
}
\examples{
library(travelpaths)
library(trackframe)
# for move2
set.seed(2025)
move2 <- sim_travel_path(5, format = "move2")
tf <- as.trackframe(data = move2)
tfb <- tf_backtransform(tf)
tfb
move2

# for sftrack
set.seed(2025)
sftrack <- sim_travel_path(5, format = "sftrack")
tf <- as.trackframe(data = sftrack)
tfb <- tf_backtransform(tf)
tfb
sftrack

# for data.frame
df <- sim_travel_path(5, format = "data.frame")
tf <- as.trackframe(data = df)
tfb <- tf_backtransform(tf)
tfb
df
}
