% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{tf_options}
\alias{tf_options}
\title{Options for col guessing in trackframe}
\usage{
tf_options(option, value)
}
\arguments{
\item{option}{a character string name of option.
"time_col", "easting_col", "northing_col", or "id_col"}

\item{value}{vector of characters with possible candidates}
}
\value{
a vector of candidates for col guessing
}
\description{
Options for col guessing in trackframe
}
\examples{
tf_options("time_col", c("t", "timestamp", "time", "time_index",
"time_col", "time_column", "tindex"))
tf_options("easting_col", c("easting", "east", "utm.easting", "easting_col",
"easting_column", "lon", "long", "longitude", "x"))
tf_options("northing_col", c("northing", "north", "utm.northing",
"northing_col", "northing_column", "lat", "latitude", "y"))
tf_options("id_col", c("track_id", "track_id", "trackid", "trackid_col",
"trackid_column", "id"))
}
