% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_frame.R
\name{track_frame}
\alias{track_frame}
\title{Creates an object to a \code{track\_frame}}
\usage{
track_frame(
  ...,
  time_col = NULL,
  easting_col = NULL,
  northing_col = NULL,
  id_col = NULL,
  utm_epsg = NULL
)
}
\arguments{
\item{...}{an object coercible to `data.frame`}

\item{time_col}{(optional) Column name of the time index column}

\item{easting_col}{(optional) Column name of the easting column}

\item{northing_col}{(optional) Column name of the northing column}

\item{id_col}{(optional) Column name of the id column}

\item{utm_epsg}{crs value for utm zone}
}
\value{
A track_frame object
}
\description{
This function creates an object of class `track_frame`.
}
\examples{
df <- data.frame(
  time_col = as.POSIXct(Sys.time() + 1:5),
  easting_col = runif(5, 0, 10),
  northing_col = runif(5, 0, 10),
  id = 1:5
)
tf <- track_frame(df, time_col = "time_col", easting_col = "easting_col",
                        northing_col = "northing_col", id_col = "id")
attributes(tf)
}
