% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_type_arrows.R
\name{type_arrows}
\alias{type_arrows}
\title{Add arrows to a plot}
\usage{
type_arrows(...)
}
\arguments{
\item{...}{arguments passed to \link[graphics:arrows]{graphics::arrows}
for example \code{length}, \code{angle}, \code{code}
length in \code{\link[graphics]{arrows}}}
}
\value{
a tinyplot_type containing corresponding draw function
}
\description{
This function adds arrows to the current (tinyplot) plot. Defines a type for tinyplots.
}
\examples{
library(trackframe)
library(tinyplot)

df_mini
tinyplot(x = df_mini$easting, y = df_mini$northing, type = "l")
tinyplot_add(
  xmin = df_mini$easting[1],
  ymin = df_mini$northing[1],
  xmax = df_mini$easting[2],
  ymax = df_mini$northing[2],
  type = type_arrows(
    length = 0.5,
    code = 2
    ),
  col = "blue",
  lty = 2,
  lwd = 2
)

}
